deeplearning-sample2/requirements.txt
ensemble_boxes==1.0.7
streamlit
matplotlib==3.4.3
torch==1.9.0
torchvision==0.10.0
requests==2.26.0
tqdm==4.62.3
pandas==1.2.3
numpy==1.20.3
opencv-python-headless
Pillow==8.3.2
PyYAML==5.4.1
scikit_learn==1.0
